pcn-helloworld: remove egress code
The egress functionality in standard cubes is more a workaround than an
actual functionality.  This support only works when ports are connected to
netdevs and the only service that makes uses of this feature so far is
pcn-iptables.

This funtionality should be not needed by most of the standard cubes, so
remove this from the helloworld example.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>