rename parameter to recirculate in send_packet_out
send_packet_out allows to sent a packet out of a cube's port and also
receive it through that port, the paremeter that controller this bevaviour
was called Direction, it was either INGRESS or EGRESS.  This naming is not
so clear as ingress and egress have a different meaning in other places.

This commit renames that parameter to recirculate, when this is false
the packet will be sent out of a port, when this this is true the packet will
be processed as it arrived to the port.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>