Remove bpf_ktime from simplebridge and use custom thread
This commit removes the bpf_ktime() helper from the Simplebridge
used to remove old entries from the filtering database.
This helper introduces a performance degradation that could be removed
by using a custom thread that updates a percpu map with the timestamp
at the same way it is done in the Iptables service.

Signed-off-by: Sebastiano Miano <mianosebastiano@gmail.com>