test standalone cube
this is an example that shows how a basic transparent cube to count packets
arriving in an interface could be implemented, not datamodel nor automatic code
generated are present in this example.

The example implements a transparent cube, it is composed of an eBPF program
that counts packets and save it in a percpu array, there is a tiny control plane
to retrieve that counter.

There is also a small rest api server to expose those statistics.