Improved performance of simplebridge during forwarding
This commit removes the bpf_map_update from the simplebridge standard
path.
In the learning phase, the filtering database is first checked with the
src MAC address of the received packet; if the entry is not present the
learning is performed, otherwise we just need to update the timestamp (or the
src port). In the latter cases, the value of the entry is directly updated
using the pointer returned from the lookup, instead of calling the update helper.
In fact, after some tests, I discovered that this function introduces a huge
performance degradation, while updating the value directly does not impact on the
overall performance.
With this commit, the multicore throughput (during forwarding) of the bridge
increased from 3.5Mpps (with 64byte packets) to ~11.5Mpps.

Signed-off-by: Sebastiano Miano <mianosebastiano@gmail.com>