services: load datamodel into file using cmake
Before this commit the datamodel has been included in the source code using
the code genereation tools.  This approach is not practical as a simple change
to it requires to regenerate the code.

This commit uses the cmake infrastructure to load the datamodel in the source
code, so a change to it will be reflected by just recompiling the service.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>