load topology at startup and save each modification to memory and file
This feature allows to load a topology from file at daemon startup,
either from the default folder or from a user selected file, like

polycubed --cubes-dump ~/Desktop/cubes.yaml

keeping the whole configuration in memory.
To enable this feature we need to use the flag --cubes-dump-enable.
After each modification in any of the topology of the cube,
the configuration in memory is updated and another thread saves it
to file updating it too.
It is possible to start the daemon with an empty topology
by using the flag --cubes-init: if the file contains a topology,
at the first update it is overwritten.

Signed-off-by: Riccardo Marchi <riccardo.marchi4@gmail.com>