API for Interface alignment.
Considering the 3 use cases:
1. the port connects to a netdev
2. transparent cube attached to a netdev
3. transparent cube attached to a port

In all 3 cases it is necessary to keep the parameters (e.g. Ip address, Mac address) between the two connected objects aligned.

This commit adds the necessary methods to the port or transparent cube to align with a netdev (case 1 and 2).
The commit also adds the skeleton for case 3, but at the moment exceptions are thrown for case 3.

Signed-off-by: francescomessina <francescomessina92@hotmail.com>